% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cello}
\alias{cello}
\title{Example Cellosaurus object}
\format{
An object of class \code{Cellosaurus} with 10 rows and 19 columns.
}
\usage{
data(cello)
}
\description{
Example Cellosaurus object
}
\examples{
data(cello)
print(cello)
}
\keyword{datasets}
