% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Cellosaurus-class}
\alias{Cellosaurus-class}
\title{Cellosaurus table of cell identifier mappings}
\value{
\code{Cellosaurus}.
}
\description{
Cellosaurus table of cell identifier mappings
}
\note{
Updated 2022-05-12.
}
