% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{Cellosaurus}
\alias{Cellosaurus}
\title{Cellosaurus table}
\usage{
Cellosaurus()
}
\value{
\code{Cellosaurus}.
}
\description{
Cellosaurus table
}
\note{
Updated 2022-05-13.
}
\examples{
## This is CPU intensive.
## > object <- Cellosaurus()
## > print(object)
}
\seealso{
\itemize{
\item ftp://ftp.expasy.org/databases/cellosaurus
\item https://github.com/calipho-sib/cellosaurus
\item ontologyIndex package
}
}
