% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CellosaurusTable-class}
\alias{CellosaurusTable-class}
\title{Cellosaurus table of cell identifier mappings}
\value{
\code{CellosaurusTable}.
}
\description{
Cellosaurus table of cell identifier mappings
}
\note{
Updated 2022-04-26.
}
