% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{CellosaurusTable}
\alias{CellosaurusTable}
\title{Cellosaurus table of cell identifier mappings}
\usage{
CellosaurusTable(x, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{x}{\code{character}.
Cellosaurus identifiers (e.g. \verb{CVCL_*}.)}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
\code{DataFrame}.
}
\description{
Cellosaurus table of cell identifier mappings
}
\note{
Updated 2020-10-01.
}
\examples{
ids <- c("CVCL_0126", "CVCL_1045")
object <- CellosaurusTable(ids)
print(object)
}
