% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerGeneFusion.R
\name{cellsPerGeneFusion}
\alias{cellsPerGeneFusion}
\title{Cells per gene fusion}
\usage{
cellsPerGeneFusion(object, minCells = 2L)
}
\arguments{
\item{object}{Object.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells per gene fusion.}
}
\value{
\code{DFrame}.
Gene fusion pairs in columns, cells in rows.
Includes additional cell line metadata on left side.
}
\description{
Cells per gene fusion
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
df <- cellsPerGeneFusion(object)
print(df[1L:5L, 1L:5L])
}
