% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerMutation.R
\name{cellsPerMutation}
\alias{cellsPerMutation}
\title{Cells per mutation}
\usage{
cellsPerMutation(object, format = c("geneName", "hgncId", "both"))
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
Logical matrix.
Genes in columns, cells in rows.
}
\description{
Cells per mutation
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)
mat <- cellsPerMutation(object, format = "geneName")
print(mat[1L:5L, 1L:5L])
}
\seealso{
\itemize{
\item \code{stringi::stri_list2matrix}.
}
}
