% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/excludeContaminatedCells-methods.R
\name{excludeContaminatedCells}
\alias{excludeContaminatedCells}
\alias{excludeContaminatedCells,Cellosaurus-method}
\title{Exclude contaminated cells}
\usage{
excludeContaminatedCells(object, ...)

\S4method{excludeContaminatedCells}{Cellosaurus}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with contaminated cells excluded (removed).
}
\description{
Exclude contaminated cells
}
\details{
\code{excludeContaminatedCells} only filters based on
\code{"Problematic cell line: Contaminated"} metadata in cell line comments. This
is less stringent than \code{excludeProblematicCells} and intentionaly keeps
cell lines that are misidentified but not contaminated.
}
\note{
Updated 2023-09-12.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
print(dim(object))
object <- excludeContaminatedCells(object)
print(dim(object))
}
\seealso{
\itemize{
\item Cellosaurus package.
}
}
