% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/excludeNonCancerCells-methods.R
\name{excludeNonCancerCells}
\alias{excludeNonCancerCells}
\alias{excludeNonCancerCells,Cellosaurus-method}
\title{Exclude non-cancer cell lines}
\usage{
excludeNonCancerCells(object, ...)

\S4method{excludeNonCancerCells}{Cellosaurus}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Exclude non-cancer cell lines
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
print(dim(object))
object <- excludeNonCancerCells(object)
print(dim(object))
}
