% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/excludeProblematicCells-methods.R
\name{excludeProblematicCells}
\alias{excludeProblematicCells}
\alias{excludeProblematicCells,Cellosaurus-method}
\title{Exclude problematic cells}
\usage{
excludeProblematicCells(object, ...)

\S4method{excludeProblematicCells}{Cellosaurus}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with problematic cells excluded (removed).
}
\description{
Exclude problematic cells
}
\details{
\code{excludeProblematicCells} filters based on \code{"Problematic cell line"} metadata
in cell line comments. This also removes contaminated cell lines, and is
more strict.
}
\note{
Updated 2023-09-12.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
print(dim(object))
object <- excludeProblematicCells(object)
print(dim(object))
}
\seealso{
\itemize{
\item Cellosaurus package.
}
}
