% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,Cellosaurus,character-method}
\title{Export data from R}
\usage{
export(object, con, ...)

\S4method{export}{Cellosaurus,character}(object, con, ...)
}
\arguments{
\item{object}{Object.}

\item{con}{\code{character(1)} or \code{connection}.
Data connection. Use \code{character(1)} to represent a file path.}

\item{...}{Passthrough arguments to \code{DFrame} method.}
}
\value{
Invisible \code{character} or \code{list} of file paths.
}
\description{
Export data from R
}
\note{
Updated 2023-09-22.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
tempdir <- AcidBase::tempdir2()
con <- file.path(tempdir, "cello.csv")
x <- export(object = object, con = con)
print(x)
AcidBase::unlink2(tempdir)
}
\seealso{
\itemize{
\item \code{BiocIO::export}.
\item \code{rio::export}.
\item \code{write.table}.
\item \code{writeLines}.
}
}
