% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/filterProblematicCells-methods.R
\name{filterContaminatedCells}
\alias{filterContaminatedCells}
\alias{filterProblematicCells}
\title{Filter problematic and/or contaminated cell lines}
\usage{
filterContaminatedCells(object, ...)

filterProblematicCells(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object, with cell lines removed.
}
\description{
Filter problematic and/or contaminated cell lines
}
\details{
\code{filterProblematicCells} filters based on \code{"Problematic cell line"} metadata
in cell line comments. This also filters contaminated cell lines, and is
more strict.

\code{filterContaminatedCells} only filters based on
\code{"Problematic cell line: Contaminated"} metadata in cell line comments. This
is less stringent than \code{filterProblematicCells} and intentionaly keeps
cell lines that are misidentified but not contaminated.
}
\note{
Updated 2023-08-23.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
x <- filterProblematicCells(object)
print(x)
y <- filterContaminatedCells(object)
print(y)
}
