% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/fusions.R
\name{fusions}
\alias{fusions}
\title{Gene fusion pairs}
\usage{
fusions(object, ...)
}
\value{
\code{CharacterList}.
}
\description{
Gene fusion pairs
}
\note{
Updated 2023-10-05
}
\examples{
data(cello)
x <- fusions(cello)
print(x)
}
