% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/geneFusions-methods.R
\name{geneFusions}
\alias{geneFusions}
\alias{geneFusions,Cellosaurus-method}
\title{Gene fusion pairs}
\usage{
geneFusions(object, ...)

\S4method{geneFusions}{Cellosaurus}(object, format = c("geneName", "hgncId", "both"))
}
\value{
\code{CharacterList}.
}
\description{
Gene fusion pairs
}
\note{
Updated 2023-10-05
}
\examples{
data(cello)
x <- geneFusions(cello)
print(x)
}
