% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells.R
\name{mapCells}
\alias{mapCells}
\title{Map cell names to identifiers using Cellosaurus}
\usage{
mapCells(x)
}
\arguments{
\item{x}{\code{character}.
Cell names or Cellosaurus identifiers.}
}
\description{
Map cell names to identifiers using Cellosaurus
}
\note{
Updated 2020-08-01.
}
\examples{
mapCells("22RV1")
}
