% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToCellosaurus.R
\name{mapCellsToCellosaurus}
\alias{mapCellsToCellosaurus}
\title{Map cell names to Cellosaurus identifiers}
\usage{
mapCellsToCellosaurus(
  x,
  organism = "Homo sapiens",
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{x}{\code{character}.
Cell names (or Cellosaurus identifiers).}
}
\value{
Named \code{character}.
User input in the names and Cellosaurus IDs in the values.
}
\description{
Map cell names to Cellosaurus identifiers
}
\note{
Updated 2020-08-03.
}
\examples{
cells <- c("22RV1", "JURKAT")
mapCellsToCellosaurus(cells)
}
