% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIXME-mappings.R
\name{mappings}
\alias{mappings}
\title{Cellosaurus identifier to cell line name mappings}
\usage{
mappings()
}
\value{
\code{DataFrame}.
Cell identifier mappings, with \code{cellId} and \code{cellName} columns.
All rows contain identifiers prefixed with \code{"CVCL_"}.
}
\description{
Cellosaurus identifier to cell line name mappings
}
\note{
Updated 2022-05-11.
}
