% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/standardizeCells-methods.R
\name{standardizeCells}
\alias{standardizeCells}
\alias{standardizeCells,Rle-method}
\alias{standardizeCells,character-method}
\title{Standardize cell line names}
\usage{
standardizeCells(object, ...)

\S4method{standardizeCells}{Rle}(object)

\S4method{standardizeCells}{character}(object)
}
\arguments{
\item{object}{\code{character}.
Cell line names.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Standardize cell line names
}
\details{
Strip all non-alphanumeric characters, remove information in parentheses or
brackets, and convert to uppercase.

Note that this function doesn't attempt to perform any mapping to the
Cellosaurus database. For that, refer to \code{mapCells} instead.
}
\note{
Updated 2023-09-21.
}
\examples{
## character ====
object <- c("22Rv1", "Jurkat", "Ramos (RA-1)")
object <- standardizeCells(object)
print(object)
}
