% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/tnbc.R
\name{tnbc}
\alias{tnbc}
\alias{tnbc,Cellosaurus-method}
\title{Triple negative breast cancer (TNBC) cell lines}
\usage{
tnbc(object, ...)

\S4method{tnbc}{Cellosaurus}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
Cellosaurus accession identifiers.
}
\description{
Triple negative breast cancer (TNBC) cell lines
}
\note{
Updated 2023-01-26.
}
\examples{
data(cello)
tnbc <- tnbc(cello)
}
