% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Cellosaurus-package}
\alias{Cellosaurus}
\alias{Cellosaurus-package}
\title{Cellosaurus}
\description{
Cellosaurus identifier mapping toolkit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acidgenomics/cellosaurus/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

}
\keyword{internal}
