% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellosaurus.R
\name{Cellosaurus}
\alias{Cellosaurus}
\title{Cellosaurus table}
\usage{
Cellosaurus()
}
\value{
\code{Cellosaurus}.
}
\description{
Cellosaurus table
}
\note{
Updated 2024-03-11.
}
\examples{
object <- Cellosaurus()
print(object)
}
\seealso{
\itemize{
\item https://www.cellosaurus.org/
\item ftp://ftp.expasy.org/databases/cellosaurus/
\item https://github.com/calipho-sib/cellosaurus/
}
}
