% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{Cellosaurus}
\alias{Cellosaurus}
\title{Cellosaurus table}
\usage{
Cellosaurus(cache = TRUE)
}
\arguments{
\item{cache}{\code{logical(1)}.
Cache files using \code{BiocFileCache}.}
}
\value{
\code{Cellosaurus}.
}
\description{
Cellosaurus table
}
\note{
Updated 2022-08-23.
}
\examples{
object <- Cellosaurus()
print(object)
}
\seealso{
\itemize{
\item ftp://ftp.expasy.org/databases/cellosaurus
\item https://github.com/calipho-sib/cellosaurus
\item ontologyIndex package
}
}
