% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{Cellosaurus}
\alias{Cellosaurus}
\title{Cellosaurus table}
\usage{
Cellosaurus()
}
\value{
\code{Cellosaurus}.
}
\description{
Cellosaurus table
}
\details{
Patient age is currently only defined in \code{cellosaurus.txt} file but
not \code{cellosaurus.obo} available from the FTP server.
}
\note{
Updated 2023-01-17.
}
\examples{
object <- Cellosaurus()
print(object)
}
\seealso{
\itemize{
\item https://www.cellosaurus.org/
\item ftp://ftp.expasy.org/databases/cellosaurus/
\item https://github.com/calipho-sib/cellosaurus/
}
}
