% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurusTable.R
\name{cellosaurusTable}
\alias{cellosaurusTable}
\title{Cellosaurus metadata table}
\usage{
cellosaurusTable(x, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{x}{\code{character}.
Cellosaurus identifiers (e.g. \verb{CVCL_*}.)}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
\code{DataFrame}.
}
\description{
Cellosaurus metadata table
}
\note{
Updated 2020-08-03.
}
\examples{
ids <- c("CVCL_0126", "CVCL_1045")
tbl <- cellosaurusTable(ids)
print(tbl)
}
