% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurusTable.R
\name{cellosaurusTable}
\alias{cellosaurusTable}
\title{Cellosaurus metadata table}
\usage{
cellosaurusTable(x, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{x}{\code{character}.
Cellosaurus identifiers (e.g. \verb{CVCL_*}.)}
}
\value{
\code{DataFrame}.
}
\description{
Cellosaurus metadata table
}
\note{
Updated 2020-08-03.
}
\examples{
ids <- c("CVCL_0126", "CVCL_1045")
tbl <- cellosaurusTable(ids)
print(tbl)
}
