% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellosaurusTable.R
\name{CellosaurusTable}
\alias{CellosaurusTable}
\title{Cellosaurus table of cell identifier mappings}
\usage{
CellosaurusTable(x)
}
\arguments{
\item{x}{\code{character}.
Cellosaurus cell line identifiers (e.g. \verb{CVCL_*}.)}
}
\value{
\code{DataFrame}.
}
\description{
Cellosaurus table of cell identifier mappings
}
\note{
Updated 2022-03-07.
}
\examples{
cells <- c("22Rv1", "Jurkat", "Ramos (RA-1)")
cells <- mapCells(cells, organism = "Homo sapiens")
print(cells)
object <- CellosaurusTable(cells)
print(object)
}
