% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellosaurusTable.R
\name{CellosaurusTable}
\alias{CellosaurusTable}
\title{Cellosaurus table of cell identifier mappings}
\usage{
CellosaurusTable(x)
}
\arguments{
\item{x}{\code{character}.
Cellosaurus cell line identifiers (e.g. \verb{CVCL_*}.)}
}
\value{
\code{DataFrame}.
}
\description{
Cellosaurus table of cell identifier mappings
}
\note{
Updated 2021-07-30.
}
\examples{
cells <- c("CVCL_0126", "CVCL_1045")
object <- CellosaurusTable(cells)
print(object)
}
