% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellsPerGeneFusion-methods.R
\name{cellsPerGeneFusion}
\alias{cellsPerGeneFusion}
\alias{cellsPerGeneFusion,Cellosaurus-method}
\title{Cells per gene fusion}
\usage{
cellsPerGeneFusion(object, ...)

\S4method{cellsPerGeneFusion}{Cellosaurus}(object, minCells = 2L)
}
\arguments{
\item{object}{Object.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells per gene fusion.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame} or \code{NULL}.
Gene fusion pairs in columns, cells in rows.
Includes additional cell line metadata on left side.
Returns \code{NULL} if no cells match or pass \code{minCells} filter.
}
\description{
Cells per gene fusion
}
\note{
Updated 2023-10-06.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
df <- cellsPerGeneFusion(object)
}
