% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerGeneFusion.R
\name{cellsPerGeneFusion}
\alias{cellsPerGeneFusion}
\title{Cells per gene fusion}
\usage{
cellsPerGeneFusion(object, format = c("geneName", "hgncId", "both"))
}
\value{
\code{matrix}.
Logical matrix.
Gene fusion pairs in columns, cells in rows.
}
\description{
Cells per gene fusion
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)
mat <- cellsPerGeneFusion(object, format = "geneName")
print(mat[1L:5L, 1L:5L])
}
