% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerMutation.R
\name{cellsPerMutation}
\alias{cellsPerMutation}
\title{Cells per mutation}
\usage{
cellsPerMutation(
  object,
  mutationFormat = c("geneName", "hgncId"),
  cellFormat = c("cellLineName", "cellosaurusId"),
  minCells = 3L
)
}
\arguments{
\item{object}{Object.}
}
\value{
\code{matrix}.
Logical matrix.
Genes in columns, cells in rows.
}
\description{
Cells per mutation
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)
mat <- cellsPerMutation(object)
print(mat[1L:5L, 1L:5L])
}
