% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cellsPerMutation-methods.R
\name{cellsPerMutation}
\alias{cellsPerMutation}
\alias{cellsPerMutation,Cellosaurus-method}
\title{Cells per mutation}
\usage{
cellsPerMutation(object, ...)

\S4method{cellsPerMutation}{Cellosaurus}(object, minCells = 2L)
}
\arguments{
\item{object}{Object.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells per mutation.}

\item{...}{Additional arguments.}
}
\value{
\code{DFrame} or \code{NULL}.
Genes in columns, cells in rows.
Includes additional cell line metadata on left side.
Returns \code{NULL} if no cells match or pass \code{minCells} filter.
}
\description{
Cells per mutation
}
\note{
Updated 2023-10-06.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
df <- cellsPerMutation(object)
print(df[1L:5L, 1L:5L])
}
