% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsPerMutation.R
\name{cellsPerMutation}
\alias{cellsPerMutation}
\title{Cells per mutation}
\usage{
cellsPerMutation(object, minCells = 2L)
}
\arguments{
\item{object}{Object.}

\item{minCells}{\code{integer(1)}.
Minimum number of cells per mutation.}
}
\value{
\code{DFrame}.
Genes in columns, cells in rows.
Includes additional cell line metadata on left side.
}
\description{
Cells per mutation
}
\note{
Updated 2023-10-06.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
mat <- cellsPerMutation(object)
print(mat[1L:5L, 1L:5L])
}
