% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentCellosaurusVersion.R
\name{currentCellosaurusVersion}
\alias{currentCellosaurusVersion}
\title{Current Cellosaurus release version}
\usage{
currentCellosaurusVersion()
}
\value{
\code{numeric_version}.
Numeric release version.
}
\description{
Current Cellosaurus release version
}
\details{
Queries the server at \code{ftp.expasy.org}.
}
\note{
Updated 2024-03-11.
}
\examples{
currentCellosaurusVersion()
}
