% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driverMutations.R
\name{driverMutations}
\alias{driverMutations}
\title{Driver mutations}
\usage{
driverMutations(object)
}
\description{
Driver mutations
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)
x <- driverMutations(cello)
print(x)
}
