% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/driverMutations-methods.R
\name{driverMutations}
\alias{driverMutations}
\alias{driverMutations,Cellosaurus-method}
\title{Driver mutations}
\usage{
driverMutations(object, ...)

\S4method{driverMutations}{Cellosaurus}(object, format = c("hgncId", "geneName", "both"))
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{CharacterList}.
}
\description{
Driver mutations
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)
x <- driverMutations(cello, format = "geneName")
print(x)
}
