% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/excludeNonHumanCells-methods.R
\name{excludeNonHumanCells}
\alias{excludeNonHumanCells}
\alias{excludeNonHumanCells,Cellosaurus-method}
\title{Exclude non-human cell lines}
\usage{
excludeNonHumanCells(object, ...)

\S4method{excludeNonHumanCells}{Cellosaurus}(object)
}
\arguments{
\item{object}{Object.}
}
\value{
Modified object, with non-human cells excluded (removed).
}
\description{
Exclude non-human cell lines
}
\note{
Updated 2023-10-06.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
print(dim(object))
object <- excludeNonHumanCells(object)
print(dim(object))
}
