% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,Cellosaurus,character,character-method}
\alias{export,Cellosaurus,character,missingOrNULL-method}
\title{Export}
\usage{
export(object, con, format, ...)

\S4method{export}{Cellosaurus,character,character}(
  object,
  con,
  format,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)

\S4method{export}{Cellosaurus,character,missingOrNULL}(
  object,
  con,
  format,
  overwrite = getOption(x = "acid.overwrite", default = TRUE),
  engine = getOption(x = "acid.export.engine", default = "base"),
  quiet = getOption(x = "acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.
An object supporting \code{dim()}, or a supported class capable of being coerced
to \code{data.frame}, to be written to disk.}

\item{con}{\code{character(1)}, \code{missing}, or \code{NULL}.
File path.
Alternatively, can leave unset and use \code{ext} and \code{dir} arguments instead.}

\item{format}{\code{character(1)}, \code{missing}, or \code{NULL}.
Output file format extension.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{engine}{\code{character(1)}.
Engine (package) to use for export.

Currently supported:
\itemize{
\item base
\item data.table
\item readr
}}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export
}
\note{
Updated 2022-08-23.
}
\section{Output file format extension}{



\code{matrix} supported arguments:
\itemize{
\item Comma separated values (CSV):
\code{"csv"}, \code{"csv.bz2"}, \code{"csv.gz"}, \code{"csv.xz"}, \code{"csv.zip"}.
\item Tab separated values (TSV):
\code{"tsv"}, \code{"tsv.bz2"}, \code{"tsv.gz"}, \code{"tsv.xz"}, \code{"tsv.zip"}.
}

\code{Matrix} (\code{sparseMatrix}) supported arguments:
\itemize{
\item MatrixMarket exchange (MTX):
\code{"mtx"}, \code{"mtx.bz2"}, \code{"mtx.gz"}, \code{"mtx.xz"}, \code{"mtx.zip"}.
}

}

\section{Row names}{



Some export utilities in R have a tendency to drop row names when writing to
disk in CSV format. For example, the \href{https://readr.tidyverse.org/}{readr} family of functions never
write row names by design. This is a \emph{really poor} default setting for
handling genomic data, which often contain gene identifiers in the row names.
Here we're performing any internal tibble coercion step to ensure row names
are always moved to a \code{"rowname"} column in the CSV export.

}

\section{Debugging}{



Note that this function currently wraps \code{readr::write_delim()} by default
for exporting \code{DataFrame}, \code{data.frame}, and \code{matrix} class objects.

}

\examples{
data(cello)

## Cellosaurus ====
object <- RangedSummarizedExperiment
con <- AcidBase::tempdir2()
x <- export(object = object, con = con)
print(x)
AcidBase::unlink2(con)
}
\seealso{
Packages:
\itemize{
\item \href{https://readr.tidyverse.org}{readr}.
\item \href{https://r-datatable.com/}{data.table}.
\item \href{https://cran.r-project.org/package=rio}{rio}.
\item \href{https://bioconductor.org/packages/rtracklayer/}{rtracklayer}.
\item \href{https://vroom.r-lib.org}{vroom}.
}

Export functions:
\itemize{
\item \code{BiocIO::export()}.
\item \code{data.table::fwrite()}.
\item \code{readr::write_delim()}.
\item \code{rio::export()}.
\item \code{rtracklayer::export()}.
\item \code{vroom::vroom_write()}.
}
}
