% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells-methods.R
\name{mapCells}
\alias{mapCells}
\alias{mapCells,character-method}
\title{Map input to cell identifiers}
\usage{
mapCells(object, ...)

\S4method{mapCells}{character}(object, organism = "Homo sapiens", BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{object}{\code{character}.
Cell names (or Cellosaurus identifiers).}

\item{...}{Additional arguments.}
}
\value{
Named \code{character}.
User input in the names and Cellosaurus IDs in the values.
}
\description{
Map input to cell identifiers
}
\note{
Updated 2020-10-01.
}
\examples{
cells <- c("22RV1", "HS729", "JURKAT")
mapCells(cells)
}
