% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapCells-methods.R
\name{mapCells}
\alias{mapCells}
\alias{mapCells,Cellosaurus-method}
\title{Map input to cell identifiers}
\usage{
mapCells(object, ...)

\S4method{mapCells}{Cellosaurus}(
  object,
  cells,
  keyType = c("cellosaurusId", "depmapId", "sangerModelId", "atccId", "cellLineName"),
  strict = FALSE
)
}
\arguments{
\item{object}{\code{character}.
Cell names (or Cellosaurus accession identifiers).}

\item{cells}{\code{character}.
Cell identifiers.}

\item{keyType}{\code{character(1)}.
Identifier format to return.}

\item{strict}{\code{logical(1)}.
\itemize{
\item \code{FALSE}, return \code{NA} for cell lines that fail to map.
\item \code{TRUE}: Error on mapping failure.
}}

\item{...}{Additional arguments.}
}
\value{
Named \code{character}.
User input in the names and Cellosaurus accession identifiers in the values.
}
\description{
Map input to cell identifiers
}
\details{
This function is designed to take input from a spreadsheet, electronic
laboratory notebook entry, or cell line provider, where the cell line names
may be inconsistent.

If you notice any mapping errors, please file a bug report.
}
\note{
Updated 2023-09-05.
}
\examples{
data(cello)

## Cellosaurus ====
## Note that the `standardizeCells` step is not required before calling
## `mapCells`. We're just performing that step here as an example where
## the cell line names change.
object <- cello
cells <- standardizeCells(head(cello[["cellLineName"]]))
print(cells)
cells <- mapCells(object = object, cells = cells)
print(cells)
}
\seealso{
\itemize{
\item \href{https://stackoverflow.com/questions/11002391/}{Matching values in list}.
}
}
