% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapCells-methods.R
\name{mapCells}
\alias{mapCells}
\alias{mapCells,character-method}
\title{Map input to cell identifiers}
\usage{
mapCells(object, ...)

\S4method{mapCells}{character}(object, organism = "Homo sapiens")
}
\arguments{
\item{object}{\code{character}.
Cell names (or Cellosaurus identifiers).}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{...}{Additional arguments.}
}
\value{
Named \code{character}.
User input in the names and Cellosaurus IDs in the values.
}
\description{
Map input to cell identifiers
}
\note{
Updated 2022-04-27.
}
\examples{
cells <- c("22Rv1", "Jurkat", "Ramos (RA-1)")
cells <- mapCells(cells)
print(cells)
}
