% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mapCells-methods.R
\name{mapCells}
\alias{mapCells}
\alias{mapCells,Cellosaurus-method}
\title{Map input to cell identifiers}
\usage{
mapCells(object, ...)

\S4method{mapCells}{Cellosaurus}(object, cells, keyType = c("cellosaurusId", "depMapId", "sangerId"))
}
\arguments{
\item{object}{\code{character}.
Cell names (or Cellosaurus identifiers).}

\item{cells}{\code{character}.
Cell identifiers.}

\item{keyType}{\code{character(1)}.
Identifier format to return.}

\item{...}{Additional arguments.}
}
\value{
Named \code{character}.
User input in the names and Cellosaurus IDs in the values.
}
\description{
Map input to cell identifiers
}
\note{
Updated 2022-06-02.
}
\examples{
## Cellosaurus ====
object <- Cellosaurus()
cells <- c("22Rv1", "Jurkat", "Ramos (RA-1)")
cells <- standardizeCells(cells)
cells <- mapCells(object = object, cells = cells)
print(cells)
}
\seealso{
\itemize{
\item \href{https://stackoverflow.com/questions/11002391/}{Matching values in list}.
}
}
