% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells.R
\name{mapCells}
\alias{mapCells}
\title{Map cell names to Cellosaurus identifiers}
\usage{
mapCells(x, organism = "Homo sapiens", BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{x}{\code{character}.
Cell names (or Cellosaurus identifiers).}

\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:register]{bpparam}.
\item \link[BiocParallel:SerialParam-class]{SerialParam}.
\item \link[BiocParallel:MulticoreParam-class]{MulticoreParam}.
}}
}
\value{
Named \code{character}.
User input in the names and Cellosaurus IDs in the values.
}
\description{
Map cell names to Cellosaurus identifiers
}
\note{
Updated 2020-10-01.
}
\examples{
cells <- c("22RV1", "HS729", "JURKAT")
mapCells(cells)
}
