% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCellsToCellosaurus.R
\name{mapCellsToCellosaurus}
\alias{mapCellsToCellosaurus}
\title{Map cell names to Cellosaurus identifiers}
\usage{
mapCellsToCellosaurus(
  x,
  organism = "Homo sapiens",
  BPPARAM = BiocParallel::bpparam()
)
}
\arguments{
\item{x}{\code{character}.
Cell names (or Cellosaurus identifiers).}
}
\description{
Map cell names to Cellosaurus identifiers
}
\note{
Updated 2020-08-01.
}
\examples{
mapCellsToCellosaurus("22RV1")
}
