% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mutations-methods.R
\name{mutations}
\alias{mutations}
\alias{mutations,Cellosaurus-method}
\title{Driver gene mutations}
\usage{
mutations(object, ...)

\S4method{mutations}{Cellosaurus}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{CharacterList}.
}
\description{
Driver gene mutations
}
\note{
Updated 2023-10-05.
}
\examples{
data(cello)

## Cellosaurus ====
object <- cello
x <- mutations(object)
print(x)
}
