% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/selectCells-methods.R
\name{selectCells}
\alias{selectCells}
\alias{selectCells,Cellosaurus-method}
\title{Select cells}
\usage{
selectCells(object, ...)

\S4method{selectCells}{Cellosaurus}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Named key value pairs that map to cell line metadata defined in \code{colData}.}
}
\value{
Modified object, subset to only contain matching cell lines.
}
\description{
Select cells
}
\details{
Only exact matching is currently supported.
}
\note{
Updated 2023-09-12.
}
\section{Supported keys}{

\itemize{
\item \code{"category"}
\item \code{"isCancer"}
\item \code{"isContaminated"}
\item \code{"isProblematic"}
\item \code{"ncbiTaxonomyId"}
\item \code{"ncitDiseaseId"}
\item \code{"ncitDiseaseName"}
\item \code{"oncotreeCode"}
\item \code{"oncotreeMainType"}
\item \code{"oncotreeName"}
\item \code{"oncotreeTissue"}
\item \code{"organism"}
\item \code{"sexOfCell"}
}
}

\examples{
data(cello)

## Cellosaurus ====
object <- cello
print(dim(object))
subset <- selectCells(
    object = object,
    category = "Cancer cell line",
    organism = "Homo sapiens",
    sexOfCell = "Female"
)
print(dim(subset))
}
\seealso{
\itemize{
\item Cellosaurus package.
\item DepMapAnalysis package.
}
}
