% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeCells.R
\name{standardizeCells}
\alias{standardizeCells}
\title{Standardize cell identifiers}
\usage{
standardizeCells(x)
}
\arguments{
\item{x}{\code{character}.
Cell line names.}
}
\value{
\code{character}.
}
\description{
Standardize cell identifiers
}
\details{
Strip all non-alphanumeric characters, remove information in
parentheses/bracket, and convert to uppercase.
}
\note{
Updated 2020-08-02.
}
\examples{
cells <- c("22Rv1", "Jurkat", "Ramos (RA-1)")
standardizeCells(c("22Rv1", "Jurkat", "Ramos (RA-1)"))
}
