% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/standardizeCells-methods.R
\name{standardizeCells}
\alias{standardizeCells}
\alias{standardizeCells,character-method}
\title{Standardize cell line names}
\usage{
standardizeCells(object, ...)

\S4method{standardizeCells}{character}(object)
}
\arguments{
\item{object}{\code{character}.
Cell line names.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Standardize cell line names
}
\details{
Strip all non-alphanumeric characters, remove information in
parentheses/brackets, and convert to uppercase.
}
\note{
Updated 2022-05-12.
}
\examples{
cells <- c("22Rv1", "Jurkat", "Ramos (RA-1)")
cells <- standardizeCells(cells)
print(cells)
}
